requirements.txt
streamlit>=1.35.0
duckdb>=0.10.0
deepface>=0.0.92
faiss-cpu>=1.8.0
git+https://github.com/openai/CLIP.git
ultralytics>=8.2.0
opencv-python-headless>=4.9.0.80
Pillow>=10.3.0
torch>=2.3.0
torchvision>=0.18.0
gdown>=5.1.0
numpy>=1.26.4
python-dotenv>=1.0.1
bcrypt>=4.1.0
requests>=2.31.0
pandas>=2.2.0
plotly>=5.22.0
